requirements.txt
godaddypy
httpx
python-dotenv